Добавляет модуляцию проекта
